Test display of checkboxes given array input
I added a test array of checkbox components, using the updated checkbox
schema. Using this test array, I updated each checkbox element to
display the required content from the test array in the appropriate
locations. This is to make sure everything works right for the eventual
checkbox data storage in NotesContext and in the future, the database.