Fix issue of body text not visible temporarily
This commit fixes the issue of the body text not being visible, for the
time being.

This does deteriorate the UI though (a medium sized note shows a scroll
bar on loading before jumping back to a better size). However, this is
meant to be a quick bug fix, and the UI will be reimproved soon in a new
feature that will be released soon.